api/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
python-multipart==0.0.20
google-generativeai==0.8.3
pillow==10.4.0
python-dotenv==1.0.1
